requirements.txt
asgiref==3.11.0
Django==5.2.8
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-extensions==4.1
pillow==12.0.0
sqlparse==0.5.3
